Sjögren's Information Application
This application provides information and brings awareness to the Auto immune disorder known as Sjögren's. Our team focused on embedding important information for people that may have or may know someone that has this disorder. We used platforms like PhoneGap and the Atom text editor, so the application also resembles a web page. The programming languages we used to design were; HTML, JavaScript and CSS.